+++ log/logtest/recorder_test.go
@@ -68,21 +68,16 @@ func TestRecorderLoggerCreatesNewStruct(t *testing.T) {
 
 func TestLoggerEnabled(t *testing.T) {
 	for _, tt := range []struct {
-		name                   string
-		options                []Option
-		ctx                    context.Context
-		buildEnabledParameters func() log.EnabledParameters
-
-		isEnabled bool
+		name          string
+		options       []Option
+		ctx           context.Context
+		enabledParams log.EnabledParameters
+		want          bool
 	}{
 		{
 			name: "the default option enables every log entry",
 			ctx:  context.Background(),
-			buildEnabledParameters: func() log.EnabledParameters {
-				return log.EnabledParameters{}
-			},
-
-			isEnabled: true,
+			want: true,
 		},
 		{
 			name: "with everything disabled",
@@ -91,17 +86,13 @@ func TestLoggerEnabled(t *testing.T) {
 					return false
 				}),
 			},
-			ctx: context.Background(),
-			buildEnabledParameters: func() log.EnabledParameters {
-				return log.EnabledParameters{}
-			},
-
-			isEnabled: false,
+			ctx:  context.Background(),
+			want: false,
 		},
 	} {
 		t.Run(tt.name, func(t *testing.T) {
-			e := NewRecorder(tt.options...).Logger("test").Enabled(tt.ctx, tt.buildEnabledParameters())
-			assert.Equal(t, tt.isEnabled, e)
+			e := NewRecorder(tt.options...).Logger("test").Enabled(tt.ctx, tt.enabledParams)
+			assert.Equal(t, tt.want, e)
 		})
 	}
 }